+++ .eslintrc.json
@@ -11,6 +11,7 @@
     },
     "parser": "babel-eslint",
     "env": {
-        "jest": true
+        "jest": true,
+        "browser": true
     }
 }
