+++ package.json
@@ -54,6 +54,7 @@
         "react-test-renderer": "^16.8.6",
         "redux": "^4.0.1",
         "redux-form": "^8.2.3",
+        "redux-thunk": "^2.3.0",
         "resolve-url-loader": "^2.3.1",
         "sass": "^1.15.2",
         "sass-loader": "^7.1.0"
