+++ resources/js/components/App.jsx
@@ -1,25 +1,44 @@
-/* global document */
 import React from 'react';
 import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware } from 'redux';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import reduxThunk from 'redux-thunk';
 import Navigation from './Navigation';
 import LandingPage from './LandingPage';
 import Footer from './Footer';
+import LoginPage from './Auth/LoginPage';
+// import RegisterPage from "./Auth/RegisterPage";
+import reducers from '../reducers';
+import { AUTHENTICATED } from '../actions';
+
+const createStoreWithMiddleware = applyMiddleware(reduxThunk)(createStore);
+const store = createStoreWithMiddleware(reducers);
+
+const user = localStorage.getItem('user');
+
+if (user) {
+    store.dispatch({ type: AUTHENTICATED });
+}
 
 const App = props => {
     const { name } = props;
 
     return (
-        <BrowserRouter>
-            <div>
-                <Navigation appName={name} />
-                <Switch>
-                    <Route exact path="/" component={LandingPage} />
-                </Switch>
-                <Footer />
-            </div>
-        </BrowserRouter>
+        <Provider store={store}>
+            <BrowserRouter>
+                <div>
+                    <Navigation appName={name} />
+                    <Switch>
+                        <Route exact path="/" component={LandingPage} />
+                        <Route path="/signin" component={LoginPage} />
+                        {/* <Route path="/secret" component={SecretPage} /> */}
+                    </Switch>
+                    <Footer />
+                </div>
+            </BrowserRouter>
+        </Provider>
     );
 };
 
@@ -28,7 +47,7 @@ App.propTypes = {
 };
 
 App.defaultProps = {
-    name: 'Lantern',
+    name: '',
 };
 
 if (document.getElementById('app')) {
